# backend/monolith/monolith-app/src/main/java/com/bestorigin/monolith/publiccontent/api/NewsFeedResponse.java
package com.bestorigin.monolith.publiccontent.api;

import java.util.List;

public record NewsFeedResponse(
        List<NewsItemResponse> items,
        NewsItemResponse featured,
        String emptyStateCode
) {
}